requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.7
Faker==37.8.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
